+++ lib/src/foundation_path_manager.dart
@@ -63,7 +63,7 @@ class FoundationPathManager extends PlatformPathManager {
   }
 
   @override
-  Future<String?> getApplicationNoBackupPath() async {
+  Future<String?> getApplicationNoBackupDirectory() async {
     final supportPath = await getApplicationSupportPath();
     if (supportPath == null) return null;