+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.6.0
+
+* Rename `getApplicationNoBackupPath` to `getApplicationNoBackupDirectory`.
+* Update `getApplicationNoBackupDirectory` to return `Future<Directory>` instead of `Future<String>` for consistency with other directory methods.
+
 # 0.5.0
 
 * initial release